collect old and new genes
In some cases, multiple genes from the target organism will match with one old gene from the template organism and all of those genes will be included in the template model. If two proteomes from different assemblies of the target organism are used to extract a draft model and we want to compare which proteome extracts the most information from the template model, instead of looking at the new genes, we should look at the number of old genes, since there may be lots of duplications in the new genes.